+++ lib/xml_parser.rb
@@ -14,4 +14,6 @@ require_relative 'xml_parser/instruction_mode'
 require_relative 'xml_parser/course_attribute'
 require_relative 'xml_parser/equivalency'
 require_relative 'xml_parser/subject'
+require_relative 'xml_parser/row'
+require_relative 'xml_parser/rows'
 Dir.glob(File.join(File.dirname(__FILE__), "xml_parser", "*.rb")) { |file| require file }
