+++ lib/xml_parser/class_rows.rb
@@ -1,19 +1,10 @@
 module PeoplesoftCourseClassData
   module XmlParser
-    class ClassRows
-      def initialize(doc)
-        self.doc = doc
-      end
-
-      def rows
-        noko_rows.map { |row| ClassRow.new(row, NAMESPACE) }
-      end
-
+    class ClassRows < Rows
       private
-      attr_accessor :doc
 
-      def noko_rows
-        doc.xpath('//ns:row', 'ns' => NAMESPACE)
+      def row_class
+        ClassRow
       end
     end
   end
