+++ lib/xml_parser/course_rows.rb
@@ -1,19 +1,10 @@
 module PeoplesoftCourseClassData
   module XmlParser
-    class CourseRows
-      def initialize(doc)
-        self.doc = doc
-      end
-
-      def rows
-        noko_rows.map { |row| CourseRow.new(row, NAMESPACE) }
-      end
-
+    class CourseRows < Rows
       private
-      attr_accessor :doc
 
-      def noko_rows
-        doc.xpath('//ns:row', 'ns' => NAMESPACE)
+      def row_class
+        CourseRow
       end
     end
   end
